+++ vpmobil/models.py
@@ -10,11 +10,11 @@ from dataclasses import dataclass
 from vpmobil.utils import prettyxml
 
 # ╭──────────────────────────────────────────────────────────────────────────────────────────╮
-# │                                         VpDay                                            │ 
+# │                                    VertretungsTag                                        │ 
 # ╰──────────────────────────────────────────────────────────────────────────────────────────╯
 
 @dataclass
-class VpDay():
+class VertretungsTag():
     """Klasse die den Vertretungsplan an einem bestimmten Tag repräsentiert.
     
     Unterstützt Subskription: 
@@ -72,7 +72,7 @@ class VpDay():
     def __repr__(self):
         return f"<Vertretungsplan vom {self.datum.strftime('%d.%m.%Y')}>"
 
-    @property     
+    @property
     def klassen(self) -> list[Klasse] | None:
         "Im Vertretungsplan hinterlegte Klassen"
         klassen: list[Klasse] = []
@@ -234,7 +234,9 @@ class Klasse():
 
 @dataclass
 class Stunde():
-    "Klasse, die eine bestimmte Unterrichtsstunde repräsentiert."
+    """Klasse, die eine bestimmte Unterrichtsstunde repräsentiert.
+    
+    Unterstützt Gleichheitsvergleich"""
 
     _data: XML.Element
         
@@ -336,6 +338,15 @@ class Stunde():
         if self.ausfall:
             return f"<Ausfall: '{self.info}'>"
         return f"<'{self.fach}' bei '{self.lehrer}' in Raum '{self.raum}'>"
+    
+    def __eq__(self, other: Stunde):
+        return (
+            self.fach == other.fach and
+            self.lehrer == other.lehrer and
+            self.raum == other.raum and
+            self.periode == other.periode and
+            self.kursnummer == other.kursnummer
+        )
 
 # ╭──────────────────────────────────────────────────────────────────────────────────────────╮
 # │                                         Kurs                                             │ 
@@ -344,6 +355,8 @@ class Stunde():
 @dataclass
 class Kurs():
     """Klasse die einen bestimmten Kurs repräsentiert.
+
+    Unterstützt Gleichheitsvergleich
     """
 
     _data: XML.Element
@@ -379,3 +392,6 @@ class Kurs():
 
     def __repr__(self) -> str:
         return f"<'{self.fach}' bei '{self.lehrer}', Gruppe '{self.gruppe or '-'}' (Kursnummer '{self.kursnummer}')>"
+    
+    def __eq__(self, other: Kurs):
+        return self.fach == other.fach and self.lehrer == other.lehrer and self.gruppe == other.gruppe and self.kursnummer == other.kursnummer